services/flexy-guard/requirements.txt
Flask==1.0.2
pymongo==3.7.2
python-dotenv==0.10.1
jsonschema==3.0.1
flask-restplus==0.12.1
netaddr==0.7.19